Improved version of lengthSort
