+++ module1/homework/lengthSort/CMakeLists.txt
@@ -41,5 +41,7 @@ enable_testing()
 add_executable(${PROJECT_NAME} main.cpp sort.cpp)
 add_executable(${PROJECT_NAME}-ut test.cpp sort.cpp)
 
+add_compile_options(${PROJECT_NAME} -Wall -Wextra -Wconversion -pedantic -Werror)
+
 target_link_libraries(${PROJECT_NAME}-ut gtest_main)
 add_test(NAME ${PROJECT_NAME}_test COMMAND ${PROJECT_NAME}-ut)
